+++ src/nl/hannahsten/texifyidea/structure/latex/LatexIncludePresentation.kt
@@ -3,8 +3,8 @@ package nl.hannahsten.texifyidea.structure.latex
 import com.intellij.navigation.ItemPresentation
 import nl.hannahsten.texifyidea.TexifyIcons
 import nl.hannahsten.texifyidea.psi.LatexCommands
-import nl.hannahsten.texifyidea.util.getIncludeCommands
 import nl.hannahsten.texifyidea.util.parser.getIncludedFiles
+import nl.hannahsten.texifyidea.util.updateAndGetIncludeCommands
 
 /**
  * @author Hannah Schellekens
@@ -14,7 +14,7 @@ class LatexIncludePresentation(labelCommand: LatexCommands) : ItemPresentation {
     private val fileName: String
 
     init {
-        if (labelCommand.name !in getIncludeCommands()) {
+        if (labelCommand.name !in updateAndGetIncludeCommands(labelCommand.project)) {
             throw IllegalArgumentException("Command $labelCommand is no include command")
         }