+++ src/nl/hannahsten/texifyidea/structure/latex/LatexStructureViewElement.kt
@@ -19,11 +19,11 @@ import nl.hannahsten.texifyidea.psi.LatexTypes
 import nl.hannahsten.texifyidea.settings.TexifySettings
 import nl.hannahsten.texifyidea.structure.bibtex.BibtexStructureViewElement
 import nl.hannahsten.texifyidea.structure.latex.SectionNumbering.DocumentClass
-import nl.hannahsten.texifyidea.util.getIncludeCommands
-import nl.hannahsten.texifyidea.util.labels.getLabelDefinitionCommands
+import nl.hannahsten.texifyidea.util.labels.getLabelDefinitionCommandsNoUpdate
 import nl.hannahsten.texifyidea.util.magic.CommandMagic
 import nl.hannahsten.texifyidea.util.parser.allCommands
 import nl.hannahsten.texifyidea.util.parser.getIncludedFiles
+import nl.hannahsten.texifyidea.util.updateAndGetIncludeCommands
 import java.util.*
 
 /**
@@ -140,7 +140,7 @@ class LatexStructureViewElement(private val element: PsiElement) : StructureView
 
     private fun addIncludes(treeElements: MutableList<TreeElement>, commands: List<LatexCommands>) {
         for (command in commands) {
-            if (command.name !in getIncludeCommands()) {
+            if (command.name !in updateAndGetIncludeCommands(command.project)) {
                 continue
             }
 
@@ -169,7 +169,7 @@ class LatexStructureViewElement(private val element: PsiElement) : StructureView
     }
 
     private fun addFromLabelingCommands(treeElements: MutableList<TreeElement>, commands: List<LatexCommands>) {
-        val labelingCommands = getLabelDefinitionCommands()
+        val labelingCommands = getLabelDefinitionCommandsNoUpdate()
         commands.filter { labelingCommands.contains(it.commandToken.text) }
             .mapNotNull { LatexStructureViewCommandElement.newCommand(it) }
             .forEach {